+++ packages/dart/npt_flutter/lib/features/profile/view/profile_header_view.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:npt_flutter/features/profile/view/pin_favorites_switch.dart';
 import 'package:npt_flutter/features/profile/widgets/profile_header_column.dart';
 import 'package:npt_flutter/features/profile_list/profile_list.dart';
 import 'package:npt_flutter/features/settings/settings.dart';
@@ -65,17 +66,29 @@ class ProfileHeaderView extends StatelessWidget {
                             children: [
                               const ProfileSelectAllBox(),
                               gapW10,
-                              ProfileHeaderColumn(
-                                title: strings.profileName,
-                                width:
-                                    SizeConfig.setProfileFieldWidthMinimalView(),
+                              Flexible(
+                                child: ProfileHeaderColumn(
+                                  title: strings.profileName,
+                                  column: SortColumn.profileName,
+                                  currentSortColumn: state.sortColumn,
+                                  sortOrder: state.sortOrder,
+                                  width:
+                                      SizeConfig.setProfileFieldWidthMinimalView(),
+                                ),
                               ),
                               gapW10,
-                              ProfileHeaderColumn(
-                                title: strings.status,
-                                width:
-                                    SizeConfig.setProfileFieldWidthMinimalView(),
+                              Flexible(
+                                child: ProfileHeaderColumn(
+                                  title: strings.status,
+                                  column: SortColumn.status,
+                                  currentSortColumn: state.sortColumn,
+                                  sortOrder: state.sortOrder,
+                                  width:
+                                      SizeConfig.setProfileFieldWidthMinimalView(),
+                                ),
                               ),
+                              gapW10,
+                              const PinFavoritesSwitch(),
                             ],
                           ),
                         ),
@@ -89,27 +102,49 @@ class ProfileHeaderView extends StatelessWidget {
                             children: [
                               const ProfileSelectAllBox(),
                               gapW10,
-                              ProfileHeaderColumn(
-                                title: strings.profileName,
-                                width: width,
-                              ),
-                              gapW10,
-                              ProfileHeaderColumn(
-                                title: strings.deviceName,
-                                width: width,
-                              ),
-                              gapW10,
-                              ProfileHeaderColumn(
-                                title: strings.serviceMapping,
-                                width: width,
-                              ),
-                              gapW10,
-                              ProfileHeaderColumn(
-                                title: strings.status,
-                                width: SizeConfig.setProfileFieldWidth(
-                                  statusField: true,
+                              Flexible(
+                                child: ProfileHeaderColumn(
+                                  title: strings.profileName,
+                                  column: SortColumn.profileName,
+                                  currentSortColumn: state.sortColumn,
+                                  sortOrder: state.sortOrder,
+                                  width: width,
                                 ),
                               ),
+                              gapW10,
+                              Flexible(
+                                child: ProfileHeaderColumn(
+                                  title: strings.deviceName,
+                                  column: SortColumn.deviceName,
+                                  currentSortColumn: state.sortColumn,
+                                  sortOrder: state.sortOrder,
+                                  width: width,
+                                ),
+                              ),
+                              gapW10,
+                              Flexible(
+                                child: ProfileHeaderColumn(
+                                  title: strings.serviceMapping,
+                                  column: SortColumn.serviceMapping,
+                                  currentSortColumn: state.sortColumn,
+                                  sortOrder: state.sortOrder,
+                                  width: width + Sizes.p25, // Extra for icon,
+                                ),
+                              ),
+                              gapW10,
+                              Flexible(
+                                child: ProfileHeaderColumn(
+                                  title: strings.status,
+                                  column: SortColumn.status,
+                                  currentSortColumn: state.sortColumn,
+                                  sortOrder: state.sortOrder,
+                                  width: SizeConfig.setProfileFieldWidth(
+                                    statusField: true,
+                                  ),
+                                ),
+                              ),
+                              gapW10,
+                              const PinFavoritesSwitch(),
                             ],
                           ),
                         ),