+++ packages/dart/npt_flutter/lib/features/profile_list/bloc/profile_list_state.dart
@@ -1,5 +1,9 @@
 part of 'profile_list_bloc.dart';
 
+enum SortColumn { profileName, deviceName, serviceMapping, status }
+
+enum SortOrder { ascending, descending }
+
 sealed class ProfileListState extends Loggable {
   const ProfileListState();
 
@@ -27,14 +31,36 @@ final class ProfileListLoading extends ProfileListState {
 
 final class ProfileListLoaded extends ProfileListState {
   final Iterable<String> profiles;
-  const ProfileListLoaded({required this.profiles});
+  final SortColumn sortColumn;
+  final SortOrder sortOrder;
+  final String searchQuery;
+  const ProfileListLoaded({
+    required this.profiles,
+    this.sortColumn = SortColumn.profileName,
+    this.sortOrder = SortOrder.ascending,
+    this.searchQuery = '',
+  });
+
+  ProfileListLoaded copyWith({
+    Iterable<String>? profiles,
+    SortColumn? sortColumn,
+    SortOrder? sortOrder,
+    String? searchQuery,
+  }) {
+    return ProfileListLoaded(
+      profiles: profiles ?? this.profiles,
+      sortColumn: sortColumn ?? this.sortColumn,
+      sortOrder: sortOrder ?? this.sortOrder,
+      // searchQuery: searchQuery ?? this.searchQuery,
+    );
+  }
 
   @override
-  List<Object> get props => [profiles];
+  List<Object> get props => [profiles, sortColumn, sortOrder, searchQuery];
 
   @override
   String toString() {
-    return 'ProfileListLoaded(profiles: $profiles)';
+    return 'ProfileListLoaded(profiles: $profiles, sortColumn: $sortColumn, sortOrder: $sortOrder, searchQuery: $searchQuery)';
   }
 }